Fix tests, add boilerplate
